src/apps/todo/serializers.py
from apps.todo.models import Task
from rest_framework import serializers


class TaskSerializer(serializers.ModelSerializer):
    user = serializers.HiddenField(default=serializers.CurrentUserDefault())

    class Meta:
        model = Task
        fields = ['text', 'user']
